Deprecate raw_config all together.
