+++ src/_pytask/hookspecs.py
@@ -77,13 +77,8 @@ def pytask_configure(
 
 
 @hookspec
-def pytask_parse_config(config: dict[str, Any], raw_config: dict[str, Any]) -> None:
-    """Parse configuration from the raw configuration that is from CLI or file.
-
-    Note that, all configuration values are transferred from the ``raw_config`` to
-    ``config`` by default.
-
-    """
+def pytask_parse_config(config: dict[str, Any]) -> None:
+    """Parse configuration that is from CLI or file."""
 
 
 @hookspec
