+++ tests/test_mark.py
@@ -49,22 +49,6 @@ def test_pytask_mark_name_starts_with_underscore():
 
 
 @pytest.mark.end_to_end
-def test_parse_markers(tmp_path):
-    toml = """
-    [tool.pytask.ini_options.markers]
-    a1 = "this is a webtest marker"
-    a2 = "this is a smoke marker"
-    """
-    tmp_path.joinpath("pyproject.toml").write_text(textwrap.dedent(toml))
-
-    session = main({"paths": tmp_path})
-
-    assert session.exit_code == ExitCode.OK
-    assert "a1" in session.config["markers"]
-    assert "a2" in session.config["markers"]
-
-
-@pytest.mark.end_to_end
 def test_markers_command(tmp_path, runner):
     toml = """
     [tool.pytask.ini_options]
